[CIR] Merge two copies of CIRGenAction.h
We have both clang/include/clang/CIRFrontendAction/CIRGenAction.h and
clang/include/clang/cir/FrontendAction/CIRGenAction.h, which is a
historical artifact. The latter is what's being upstreamed, so merge the
former into it to avoid any confusion.
